[24.12.12] 메인페이지 전사게시판 데이터 연결
