+++ semi-project/src/main/resources/templates/main/main.html
@@ -10,6 +10,8 @@
     <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
     <script src="/main/js/user-info.js"></script>
     <link rel="stylesheet" href="/main/css/user-info.css">
+    <script src="/main/js/slide-bar.js"></script>
+    <link rel="stylesheet" href="/main/css/slide-bar.css">
 </head>
 <body>
 <header>
@@ -58,12 +60,9 @@
     <div class="change-screen">
         <div id="change-screen-content">
 
-            <div class="slide-board">
-
-            </div>
+            <div class="slide-board"></div>
 
             <div class="main-form1">
-
                 <div class="main-form2">
                     <div id="user-info">
                         <div class="profile-container">
